+++ bld/f77/wfc/h/implicit.h
@@ -2,6 +2,7 @@
 *
 *                            Open Watcom Project
 *
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -29,4 +30,4 @@
 ****************************************************************************/
 
 
-extern bool     LenSpec( TYPE typ, uint *size_ptr );
+extern bool     LenSpec( TYPE typ, size_t *size_ptr );
